requirements.txt
streamlit
pandas
snowflake-snowpark-python
python-dotenv
pyarrow<19.0.0
schedule
